Implement Cantillon support and refactor JS side (#124)
* User is asked for pin at app startup. Cache it globally and only ask for it again upon sending extrinsics. Closes #111 
* Dart: Integrated `TrustedGetters`. Closes #108 
* Implemented `TrustedCalls`. Closes #109 
* Dart: `ApiEnointer` now has modular child apis `ApiNoTee` and `ApiTeeProxy`
* Add option to register with `ProofOfAttendance` of last meetup. Closes #10 
* Close open js-channel subscriptions, when launching a webView. Closes #102 
* Fix #126 by using the worker from @encointer
* JS: use @encointer packages for the worker and some other stuff. Closes #128 
* JS: Added `SubstrateeRegistry` rpc calls.
* JS: Restructured tests. Moved them to a separate directory.

Tested with a successful meetup on local Cantillon.

+++ lib/store/encointer/types/location.dart
@@ -9,17 +9,14 @@ part 'location.g.dart';
 class Location {
   Location(this.lon, this.lat);
 
-  final BigInt lon;
-  final BigInt lat;
+  final int lon;
+  final int lat;
 
   @override
   String toString() {
     return jsonEncode(this);
   }
 
-
-  factory Location.fromJson(Map<String, dynamic> json) =>
-      _$LocationFromJson(json);
-  Map<String, dynamic> toJson() =>
-      _$LocationToJson(this);
+  factory Location.fromJson(Map<String, dynamic> json) => _$LocationFromJson(json);
+  Map<String, dynamic> toJson() => _$LocationToJson(this);
 }
